Moves container traits from ygm/container/detail up a directory
